+++ Modules/Ruleset.ProgramInfo/Private/Initialize-Table.ps1
@@ -36,7 +36,7 @@ as users or groups and their SID for which given firewall rule applies
 This method is primarily used to reset the table
 Each entry in the table also has an ID to help choosing entries by ID
 
-.PARAMETER TableName
+.PARAMETER Name
 Table name
 
 .EXAMPLE
@@ -59,7 +59,7 @@ function Initialize-Table
 	[OutputType([void])]
 	param (
 		[Parameter()]
-		[string] $TableName = "InstallationTable"
+		[string] $Name = "InstallTable"
 	)
 
 	Write-Debug -Message "[$($MyInvocation.InvocationName)] params($($PSBoundParameters.Values))"
@@ -68,31 +68,31 @@ function Initialize-Table
 	if ($Develop)
 	{
 		Write-Debug -Message "[$($MyInvocation.InvocationName)] resetting global installation table"
-		Set-Variable -Name InstallTable -Scope Global -Value (New-Object -TypeName System.Data.DataTable $TableName)
+		Set-Variable -Name $Name -Scope Global -Value (New-Object -TypeName System.Data.DataTable $Name)
 	}
 	else
 	{
 		Write-Verbose -Message "[$($MyInvocation.InvocationName)] resetting local installation table"
-		Set-Variable -Name InstallTable -Scope Script -Value (New-Object -TypeName System.Data.DataTable $TableName)
+		Set-Variable -Name $Name -Scope Script -Value (New-Object -TypeName System.Data.DataTable $Name)
 	}
 
 	Set-Variable -Name RowIndex -Scope Script -Value 0
 
 	# Define Columns
 	$ColumnID = New-Object -TypeName System.Data.DataColumn ID, ([int32])
-	$ColumnSID = New-Object -TypeName System.Data.DataColumn SID, ([string])
 	$ColumnUser = New-Object -TypeName System.Data.DataColumn User, ([string])
 	$ColumnGroup = New-Object -TypeName System.Data.DataColumn Group, ([string])
-	$ColumnAccount = New-Object -TypeName System.Data.DataColumn Account, ([string])
-	$ColumnComputer = New-Object -TypeName System.Data.DataColumn Computer, ([string])
+	$ColumnDomain = New-Object -TypeName System.Data.DataColumn Domain, ([string])
+	$ColumnPrincipal = New-Object -TypeName System.Data.DataColumn Principal, ([string])
+	$ColumnSID = New-Object -TypeName System.Data.DataColumn SID, ([string])
 	$ColumnInstallLocation = New-Object -TypeName System.Data.DataColumn InstallLocation, ([string])
 
 	# Add the Columns
 	$InstallTable.Columns.Add($ColumnID)
-	$InstallTable.Columns.Add($ColumnSID)
 	$InstallTable.Columns.Add($ColumnUser)
 	$InstallTable.Columns.Add($ColumnGroup)
-	$InstallTable.Columns.Add($ColumnAccount)
-	$InstallTable.Columns.Add($ColumnComputer)
+	$InstallTable.Columns.Add($ColumnDomain)
+	$InstallTable.Columns.Add($ColumnPrincipal)
+	$InstallTable.Columns.Add($ColumnSID)
 	$InstallTable.Columns.Add($ColumnInstallLocation)
 }